pyproject.toml
[project]
name = "resume-screening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.1",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
